Unify the launch way of lexical analysis (#234)
* lac and waybill use int64 for windows compatibility

* replace spawn

* optimize readme

* use --device gpu

+++ examples/information_extraction/waybill_ie/run_bigru_crf.py
@@ -134,9 +134,9 @@ if __name__ == '__main__':
     test_ds.map(convert_example)
 
     batchify_fn = lambda samples, fn=Tuple(
-        Pad(axis=0, pad_val=word_vocab.get('OOV')),  # token_ids
-        Stack(),  # seq_len
-        Pad(axis=0, pad_val=label_vocab.get('O'))  # label_ids
+        Pad(axis=0, pad_val=word_vocab.get('OOV'), dtype='int64'),  # token_ids
+        Stack(dtype='int64'),  # seq_len
+        Pad(axis=0, pad_val=label_vocab.get('O'), dtype='int64')  # label_ids
     ): fn(samples)
 
     train_loader = paddle.io.DataLoader(
